Get station metadata

Returns metadata about measurement stations in the EA network.

Usage

ea_station_list(
  station_guid,
  sample_of,
  status_label,
  observed_property,
  open_from,
  open_to,
  long,
  lat,
  easting,
  northing,
  dist,
  ...
)

Arguments

station_guid

Character. Station globally unique identifer. Only one station can be provided at once.

sample_of

Character. River name (e.g. "River Thames"). Note that rivers are named inconsistently, so it may be necessary to supply multiple versions of river names.

status_label

Character. One of 'Active', 'Suspended' or 'Closed'

observed_property

Character.

open_from

Date or character formatted "YYYY-MM-DD"

open_to

Date or character formatted "YYYY-MM-DD"

long

Numeric. The longitude of the centre of the search area with radius dist.

lat

Numeric. The latitude of the centre of the search area with radius dist.

easting

Numeric. The easting of the centre of the search area with radius dist.

northing

Numeric. The northing of the centre of the search area with radius dist.

dist

Numeric. Distance in kilometres from the point specified by long and lat or easting and northing.

...

Additional arguments. None implemented.

Value

A tibble.

Measured variables

The argument observed_property allows you to select stations based on the variables they measure. Here are the available options:

waterFlow

waterLevel

rainfall

groundwaterLevel

dissolved-oxygen

fdom

bga

turbidity

chlorophyll

conductivity

temperature

ammonium

nitrate

ph

Examples

if (FALSE) { # \dontrun{ 
# Get all stations that collect river discharge 
x <- ea_station_list(observed_property = "waterFlow")

# Get metadata for a certain station
id <- x$stationGuid[1]
y <- ea_station_list(station_guid = id)

# Get metadata for all stations on the Thames
thames_nms <- c("River Thames", "RIVER THAMES", "THAMES")
thames_stations <- ea_station_list(sample_of = thames_nms)

# Get discharge stations on the River Thames within 20km of Abingdon
abingdon_stations <- ea_station_list(
  sample_of = thames_nms, observed_property = "waterFlow",
  lat = 51.6708, long = -1.2880, dist = 20
)
} # }