Get time series data for one or more stations

ea_timeseries_values() is used to retrieve timeseries data in one of two ways. Firstly, you may supply the measure argument, which is a string that uniquely identifies the measurement and is provided in the notation column in the tibble returned by ea_timeseries_list(). Alternatively, you can provide station_guid, observed_property, value_type and period to uniquely define the measurement.

Usage

ea_timeseries_values(
  measure,
  station_guid,
  observed_property,
  value_type,
  period,
  date,
  min_date,
  max_date,
  earliest = FALSE,
  latest = FALSE,
  ...
)

Arguments

measure

Character. String that uniquely identifies the measurement. This is provided in the notation column of the tibble returned by ea_timeseries_list().

station_guid

Character. Station global unique identifier. Not used if measure is supplied.

observed_property

Character. Observation property. Not used if measure is supplied.

value_type

Character. Value type. Not used if measure is supplied.

period

Integer. Period between readings in seconds. Not used if measure is supplied.

date

Date or character formatted "YYYY-MM-DD"

min_date

Date or character formatted "YYYY-MM-DD"

max_date

Date or character formatted "YYYY-MM-DD"

earliest

Logical. If TRUE, will return the earliest measurements on record.

latest

Logical. If TRUE, will return the latest measurements on record.

...

Additional arguments. None implemented.

Value

A tibble.

Examples

if (FALSE) { # \dontrun{

# Get discharge stations on the River Thames within 5km of Abingdon
abingdon_stations <- ea_station_list(
  sample_of = "River Thames", observed_property = "waterFlow",
  lat = 51.6708, long = -1.2880, dist = 5
)
guid <- abingdon_stations$stationGuid
tslist <- ea_timeseries_list(guid)

# Filter to get daily mean flow 
measure <- tslist |> 
  filter(period == 86400 & valueType == "mean") |> 
  pull(notation)
min_date <- Sys.Date() - lubridate::days(31)
ts <- ea_timeseries_values(measure, min_date = min_date)

# Alternatively: 
ts <- ea_timeseries_values(
  station_guid = guid, observed_property = "waterFlow",
  value_type = "mean", period = 86400, min_date = min_date
)
} # }